firebase backend file refactor
