+++ functions/util/firebaseConfig.js
@@ -1,0 +1,9 @@
+module.exports = {
+  apiKey: "AIzaSyDH_hiOBxaDyXSyZ6EaXtLtbx2Cv9tblHI",
+  authDomain: "quickstart-1561998550467.firebaseapp.com",
+  databaseURL: "https://quickstart-1561998550467.firebaseio.com",
+  projectId: "quickstart-1561998550467",
+  storageBucket: "quickstart-1561998550467.appspot.com",
+  messagingSenderId: "1091712667896",
+  appId: "1:1091712667896:web:7f3c2f9dd78115b3"
+}
